refactor: 피드백 반영

+++ src/main/java/qna/domain/Question.java
@@ -8,6 +8,7 @@ import javax.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
 import org.springframework.util.CollectionUtils;
+import qna.CannotDeleteException;
 
 @Entity
 public class Question extends AbstractEntity {
@@ -99,15 +100,25 @@ public class Question extends AbstractEntity {
         return answers.stream().allMatch(answer -> answer.isOwner(loginUser));
     }
 
-    public List<DeleteHistory> delete() {
+    public List<DeleteHistory> deleteBy(User loginUser) throws CannotDeleteException {
+        if (!isOwner(loginUser)) {
+            throw new CannotDeleteException("질문을 삭제할 권한이 없습니다.");
+        }
+
+        if (!isDeletableBy(loginUser)) {
+            throw new CannotDeleteException("다른 사람이 쓴 답변이 있어 삭제할 수 없습니다.");
+        }
+
         List<DeleteHistory> deleteHistories = deleteAllAnswer();
         deleted = true;
         deleteHistories.add(new DeleteHistory(ContentType.QUESTION, getId(), writer, LocalDateTime.now()));
         return deleteHistories;
     }
 
-    public List<DeleteHistory> deleteAllAnswer() {
-        return answers.stream().map(Answer::delete).collect(Collectors.toList());
+    private List<DeleteHistory> deleteAllAnswer() {
+        return answers.stream()
+            .map(Answer::delete)
+            .collect(Collectors.toList());
     }
 
     @Override
